requirements.txt
Flask==2.1.1
gunicorn==20.1.0
google-cloud-bigquery==2.31.0
python-dotenv
requests
pandas
numpy<2.0
pyyaml
cloudops
google-api-python-client
google-cloud
google-cloud-vision
google-cloud-storage
google-cloud-speech
google-cloud-logging
pydantic
pandas-gbq
Werkzeug==2.0.3